requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.14.0
keras==2.14.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
joblib==1.3.2
aiohttp==3.9.1
